pyproject.toml
[tool.poetry]
name = "synomail"
version = "0.1.0"
description = ""
authors = ["Antonio Alonso"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
flask-migrate = "^4.0.5"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.2"
waitress = "^2.1.2"
bootstrap-flask = "^2.3.0"
cryptography = "^41.0.4"
flask-wtf = "^1.1.1"
synochat = "^1.0.4"
docx-hitalent = "^0.8.51"
wtforms-sqlalchemy = "^0.3"
synology-drive-api = {git = "https://github.com/alloant/synology-drive-api.git", rev = "v0.1.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
